+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2024-06-16
+
+- [Issue #26](https://github.com/itsallcode/openfasttrace-gradle/issues/26)
+  - Added option `failBuild` that lets the build fail when it finds defects
+  - **Breaking Change:** `failBuild` is set to `true` by default. To keep the previous behaviour use `failBuild = false` in your build.
+
 ## [2.0.0] - 2024-06-13
 
 - [PR #44](https://github.com/itsallcode/openfasttrace-gradle/pull/35)